# docs/02-architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: Forge Platform API
  version: 1.0.0
  description: >-
    REST contract consumed by the Forge SPA (catalog, approval queue, trace
    viewer). Consistent with the data model (data-model.md) and the agent DNA
    contract (dna-schema.json). Governance is structural: publishing is
    eval-gated (409 when the gate has not passed), approvals expire server-side
    and there is no extend/auto-approve operation, and a role header
    demonstrates segregation of duties (NFR-5). This is a demonstration
    contract, not a production auth design.
servers:
  - url: /api/v1
...
paths:
...
  /approvals:
    get:
      tags: [Approvals]
      summary: List approvals (defaults to pending)
      description: >-
        The queue as an approver works it, oldest first, each item carrying the
        proposed action **and the evidence to decide it** (FR-E1) so a decision
        needs no second tab and no second request. Needs `read`.

        Expiry is applied before the list is built, so nothing returned as
        pending has already lapsed — and asking for the queue is one of the
        moments a lapsed approval is canceled.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, granted, rejected, expired]
            default: pending
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /approvals/report:
    get:
      tags: [Approvals]
      summary: Autonomy-promotion report — approval rates per action category (read-only)
      description: >-
        What the approvers actually did, per agent version and tool — the pair a
        DNA grant names (FR-E5). Answers Rosa's approval-fatigue risk with
        evidence: a category with a long run of grants and no rejections is
        named as a candidate for an autonomy upgrade.

        **Read-only, and there is no endpoint that applies it.** Autonomy is a
        property of a published DNA document, so raising it means authoring a
        new version and putting it through the eval gate — never a statistic
        crossing a line. Needs `read`.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: One row per action category, busiest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AutonomyCandidate' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /approvals/{approvalId}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      tags: [Approvals]
      summary: Get one approval with its evidence
      description: >-
        Needs `read`. Expiry is enforced here too: no path serves a lapsed
        approval as though it were still live.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /approvals/{approvalId}/approve:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      tags: [Approvals]
      summary: Approve a pending action (resumes the run)
      description: >-
        Releases **exactly this action** — the call the run parked, with the
        arguments the tool gateway already validated. The request body carries a
        note and no arguments, so there is no shape of request that approves one
        action and runs another (FR-E2). The approval is written `granted`
        before anything executes; the run then continues through the same tool
        gateway and reaches a terminal state of its own.

        Needs `approval.decide`, which only the approver role holds and which no
        role holding `agent.configure` may ever hold (NFR-5). A refused attempt
        is recorded as a `governance.permission_denied` event. Fails with 409 if
        the approval is not pending — including one whose deadline has passed.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Decision' }
      responses:
        '200':
          description: Approved; the run resumed and executed the released action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Approval is no longer pending (already decided or expired)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: approval_not_pending
                message: >-
                  approval ... is expired; only a pending approval can be
                  decided. Its deadline passed and the run was canceled — expiry
                  never approves, and no operation extends it
  /approvals/{approvalId}/reject:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      tags: [Approvals]
      summary: Reject a pending action (cancels the run)
      description: >-
        Refuses the action and cancels its run; nothing is executed, and the
        refusal is recorded with the actor, the timestamp, and the optional note
        (FR-E4). Needs `approval.decide`.

        **These four operations are the whole surface.** There is deliberately
        no extend, no snooze, no bulk-approve and no auto-approve anywhere in
        this API: expiry is enforced server-side and always cancels, never
        approves (FR-E3, fail-closed). The absence is the control, and
        `tests/test_approvals.py` asserts it against the served contract.
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Decision' }
      responses:
        '200':
          description: Rejected; the run was canceled and nothing ran
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Approval is no longer pending
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  parameters:
    RoleHeader:
      name: X-Forge-Role
      in: header
      required: true
      description: >-
        Demonstration role for segregation of duties (NFR-5). configurator
        holds read, agent.configure, agent.publish, agent.suspend and
        run.start; approver holds read and approval.decide; viewer holds read
        only; admin holds read, agent.suspend and agent.resume — the
        containment operator, deliberately not a superuser. No role holds both
        agent.configure/agent.publish and approval.decide, and none holds both
        agent.configure/agent.publish and agent.resume: the person who decides
        what an agent may do is never the person who approves what it proposes,
        and whoever built or shipped an agent can never override the breaker
        that contained it. The build refuses to start if either pairing ever
        stops being true. A role lacking the permission for an operation gets
        403, and the refusal is recorded as a governance.permission_denied
        event. Not authentication.
      schema:
        type: string
        enum: [configurator, approver, viewer, admin]
...
    ApprovalId:
      name: approvalId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: The X-Forge-Role is not permitted to perform this operation (NFR-5)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine-readable error code }
        message: { type: string, description: Human-readable summary }
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (e.g. validation errors)
...
    Approval:
      type: object
      required:
        - id
        - tenant_id
        - run_id
        - run_status
        - status
        - proposed_action
        - evidence
        - expires_at
        - seconds_remaining
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string }
        run_id: { type: string, format: uuid }
        run_status:
          type: string
          enum: [running, awaiting_approval, completed, escalated, canceled, error]
          description: >-
            State of the run this approval is holding. `awaiting_approval` while
            pending; after a decision it is whatever the run reached — canceled
            for a rejection or an expiry, and the loop's own outcome for a grant.
        status: { type: string, enum: [pending, granted, rejected, expired] }
        proposed_action: { $ref: '#/components/schemas/ProposedAction' }
        evidence: { $ref: '#/components/schemas/ApprovalEvidence' }
        why_approval_required:
          type: string
          description: >-
            The plain-language sentence the platform recorded when it parked the
            action (the `approval_required` explanation). Served rather than
            written into each client, so the words an approver reads are the
            words the platform acted on.
        expires_at:
          type: string
          format: date-time
          description: >-
            Server-side deadline, written once when the action was parked from
            the agent's `guardrails.approval_sla_seconds` (or the platform
            default). On expiry the run is **canceled**, never approved, and no
            operation in this API moves this value (FR-E3).
        seconds_remaining:
          type: integer
          description: Whole seconds left before the deadline; 0 once it has passed
        decision: { type: [string, 'null'], enum: [approve, reject, null] }
        decided_by:
          type: [string, 'null']
          description: >-
            `role:<role>` for a human decision, `system` for an expiry — which is
            the absence of a decision, so `decision` stays null there.
        decided_at:
          type: [string, 'null']
          format: date-time
          description: >-
            When it was decided; for an expiry, the deadline itself — authority
            lapsed when the clock said so, not when the platform next looked.
        note: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
    ProposedAction:
      type: object
      description: >-
        The action the run parked, exactly as the gateway validated it. This *is*
        the scope of the approval: one tool, one set of arguments. Approving it
        authorises this and nothing else — not the same tool again, and not the
        same tool with a different amount (FR-E2).
      required: [tool_invocation_id, tool_ref, autonomy, status]
      properties:
        tool_invocation_id: { type: string, format: uuid }
        tool_ref: { type: string, description: slug@semver, as granted in the DNA }
        autonomy: { type: string, enum: [autonomous, requires_approval, forbidden] }
        args: { type: object, additionalProperties: true }
        status:
          type: string
          description: Always `validated` while pending — checked, and deliberately not run
    ApprovalEvidence:
      type: object
      description: >-
        Everything the agent gathered before it asked for a human, verbatim from
        the run's append-only log (FR-E1). Served with the queue rather than
        behind a second request, because the requirement is a decision in under
        a minute and a round trip per invoice is not that.
      required: [agent, run_input, observations, rule_ids]
      properties:
        agent: { type: string, description: slug@semver of the version that proposed it }
        agent_description: { type: [string, 'null'] }
        run_input:
          type: object
          additionalProperties: true
          description: The trigger payload the run started from
        observations:
          type: array
          items: { $ref: '#/components/schemas/ApprovalObservation' }
        rule_ids:
          type: array
          items: { type: string }
          description: >-
            Governed rule ids present in what the agent gathered — the rules in
            play, so "which rule fired" is on the screen rather than inside a
            payload. A reading of the evidence, not a decision: the agent has not
            made one. The authoritative citations are the ones a decision carries
            (R-092).
    ApprovalObservation:
      type: object
      description: One tool call the agent executed before it asked for a human
      required: [tool_invocation_id, tool_ref, tool_name]
      properties:
        tool_invocation_id: { type: string, format: uuid }
        tool_ref: { type: string }
        tool_name: { type: string }
        args: { type: object, additionalProperties: true }
        result: { type: object, additionalProperties: true }
    AutonomyCandidate:
      type: object
      description: >-
        One action category in the autonomy-promotion report (FR-E5): one agent
        version, one tool. Read-only — nothing in this API applies a promotion.
      required: [agent, agent_version_id, tool_ref, granted, rejected, expired, candidate]
      properties:
        agent: { type: string, description: slug@semver }
        agent_version_id: { type: string, format: uuid }
        tool_ref: { type: string }
        pending: { type: integer }
        granted: { type: integer }
        rejected: { type: integer }
        expired:
          type: integer
          description: >-
            Approvals nobody answered. Each one canceled its run; none of them
            counts as consent.
        decided: { type: integer, description: granted + rejected — expiry is not a decision }
        approval_rate:
          type: [number, 'null']
          description: granted / decided; null when nothing has been decided
        candidate:
          type: boolean
          description: >-
            Enough decisions to mean something, no rejection ever, and a grant
            rate above the threshold. A suggestion for a reviewer, applied only
            by publishing a new DNA version through its eval gate.
        recommendation: { type: string }
        fatigue_note:
          type: [string, 'null']
          description: Set when approvals in this category expired unanswered
    Decision:
      type: object
      additionalProperties: false
      description: >-
        The body of approve and reject: a note, and nothing else. There are
        deliberately **no arguments** here — what is released is the action the
        run already parked (FR-E2) — and no `extend` field, because no operation
        anywhere in this API moves a deadline (FR-E3).
      properties:
        note: { type: string, description: Optional reviewer note recorded with the decision }
...